Move use-links to /store
